// apps/client/src/app/core/swagger/model/report.ts
/**
 * NgReports
 * NG Reports API
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Comment } from './comment';
import { NgReportsAttachmentImage } from './ngReportsAttachmentImage';
import { NgReportsAuthUserPartial } from './ngReportsAuthUserPartial';
import { NgReportsEnvironment } from './ngReportsEnvironment';
import { NgReportsLogBase } from './ngReportsLogBase';

export interface Report { 
    dataIsFromAuthService: boolean;
    user?: NgReportsAuthUserPartial;
    title: string;
    details: string;
    logs?: Array<NgReportsLogBase>;
    formData?: any;
    attachments?: Array<NgReportsAttachmentImage>;
    environment: NgReportsEnvironment;
    timestamp: number;
    severity?: Report.SeverityEnum;
    summary?: string;
    fixed?: boolean;
    comments: Array<Array<Comment>>;
}
export namespace Report {
    export type SeverityEnum = 0 | 1 | 2 | 3;
    export const SeverityEnum = {
        NUMBER_0: 0 as SeverityEnum,
        NUMBER_1: 1 as SeverityEnum,
        NUMBER_2: 2 as SeverityEnum,
        NUMBER_3: 3 as SeverityEnum
    };
}